# BizzyFrontend/contracts/backend.openapi.json
{
  "components": {
    "schemas": {
      "AgentResponse": {
        "properties": {
          "agent": {
            "title": "Agent",
            "type": "string"
          },
          "confidence": {
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Confidence",
            "type": "number"
          },
          "evidence": {
            "items": {
              "$ref": "#/components/schemas/Evidence"
            },
            "title": "Evidence",
            "type": "array"
          },
          "recommended_actions": {
            "items": {
              "$ref": "#/components/schemas/RecommendedAction"
            },
            "title": "Recommended Actions",
            "type": "array"
          },
          "status": {
            "$ref": "#/components/schemas/AgentStatus"
          },
          "summary": {
            "title": "Summary",
            "type": "string"
          }
        },
        "required": [
          "agent",
          "status",
          "summary",
          "confidence"
        ],
        "title": "AgentResponse",
        "type": "object"
      },
      "AgentStatus": {
        "enum": [
          "success",
          "partial",
          "failed"
        ],
        "title": "AgentStatus",
        "type": "string"
      },
      "AlertSeverity": {
        "enum": [
          "critical",
          "warning",
          "info"
        ],
        "title": "AlertSeverity",
        "type": "string"
      },
      "BusinessAlert": {
        "properties": {
          "agent": {
            "title": "Agent",
            "type": "string"
          },
          "alert_id": {
            "title": "Alert Id",
            "type": "string"
          },
          "alert_type": {
            "title": "Alert Type",
            "type": "string"
          },
          "as_of": {
            "format": "date",
            "title": "As Of",
            "type": "string"
          },
          "current_value": {
            "title": "Current Value"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "metric": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Metric"
          },
          "product_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Product Id"
          },
          "recommended_action": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RecommendedAction"
              },
              {
                "type": "null"
              }
            ]
          },
          "severity": {
            "$ref": "#/components/schemas/AlertSeverity"
          },
          "threshold": {
            "title": "Threshold"
          },
          "title": {
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "alert_id",
          "alert_type",
          "severity",
          "agent",
          "title",
          "message",
          "as_of"
        ],
        "title": "BusinessAlert",
        "type": "object"
      },
      "DecisionRequest": {
        "properties": {
          "reason": {
            "maxLength": 1000,
            "minLength": 1,
            "title": "Reason",
            "type": "string"
          }
        },
        "required": [
          "reason"
        ],
        "title": "DecisionRequest",
        "type": "object"
      },
      "Evidence": {
        "properties": {
          "dimensions": {
            "additionalProperties": {
              "type": "string"
            },
            "title": "Dimensions",
            "type": "object"
          },
          "metric": {
            "title": "Metric",
            "type": "string"
          },
          "period": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Period"
          },
          "source": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source"
          },
          "unit": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Unit"
          },
          "value": {
            "title": "Value"
          }
        },
        "required": [
          "metric",
          "value"
        ],
        "title": "Evidence",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "QueryRequest": {
        "properties": {
          "language": {
            "default": "en",
            "maxLength": 16,
            "minLength": 2,
            "title": "Language",
            "type": "string"
          },
          "question": {
            "maxLength": 4000,
            "minLength": 1,
            "title": "Question",
            "type": "string"
          },
          "user": {
            "default": "owner",
            "title": "User",
            "type": "string"
          }
        },
        "required": [
          "question"
        ],
        "title": "QueryRequest",
        "type": "object"
      },
      "QueryResponse": {
        "properties": {
          "advisor_result": {
            "$ref": "#/components/schemas/AgentResponse"
          },
          "approval_required": {
            "title": "Approval Required",
            "type": "boolean"
          },
          "guard_decision": {
            "title": "Guard Decision",
            "type": "string"
          },
          "invoked_agents": {
            "items": {
              "type": "string"
            },
            "title": "Invoked Agents",
            "type": "array"
          },
          "specialist_results": {
            "items": {
              "$ref": "#/components/schemas/AgentResponse"
            },
            "title": "Specialist Results",
            "type": "array"
          },
          "workflow_id": {
            "title": "Workflow Id",
            "type": "string"
          }
        },
        "required": [
          "workflow_id",
          "invoked_agents",
          "specialist_results",
          "advisor_result",
          "guard_decision",
          "approval_required"
        ],
        "title": "QueryResponse",
        "type": "object"
      },
      "RecommendedAction": {
        "properties": {
          "expected_impact": {
            "additionalProperties": true,
            "title": "Expected Impact",
            "type": "object"
          },
          "parameters": {
            "additionalProperties": true,
            "title": "Parameters",
            "type": "object"
          },
          "reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          },
          "risk_level": {
            "$ref": "#/components/schemas/RiskLevel",
            "default": "GREEN"
          },
          "type": {
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "type"
        ],
        "title": "RecommendedAction",
        "type": "object"
      },
      "RiskLevel": {
        "enum": [
          "GREEN",
          "AMBER",
          "RED"
        ],
        "title": "RiskLevel",
        "type": "string"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "BizzyBee Backend",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/actions/{action_id}/approve": {
      "post": {
        "operationId": "approve_api_v1_actions__action_id__approve_post",
        "parameters": [
          {
            "in": "path",
            "name": "action_id",
            "required": true,
            "schema": {
              "title": "Action Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DecisionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Approve Api V1 Actions  Action Id  Approve Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Approve",
        "tags": [
          "bizzybee"
        ]
      }
    },
    "/api/v1/actions/{action_id}/reject": {
      "post": {
        "operationId": "reject_api_v1_actions__action_id__reject_post",
        "parameters": [
          {
            "in": "path",
            "name": "action_id",
            "required": true,
            "schema": {
              "title": "Action Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DecisionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Reject Api V1 Actions  Action Id  Reject Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Reject",
        "tags": [
          "bizzybee"
        ]
      }
    },
    "/api/v1/audit/{workflow_id}": {
      "get": {
        "operationId": "audit_event_api_v1_audit__workflow_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "workflow_id",
            "required": true,
            "schema": {
              "title": "Workflow Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Audit Event Api V1 Audit  Workflow Id  Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Audit Event",
        "tags": [
          "bizzybee"
        ]
      }
    },
...
    "/api/v1/finance/summary": {
      "get": {
        "operationId": "finance_summary_api_v1_finance_summary_get",
        "parameters": [
          {
            "in": "query",
            "name": "language",
            "required": false,
            "schema": {
              "default": "en",
              "maxLength": 16,
              "minLength": 2,
              "title": "Language",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "as_of",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "As Of"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Finance Summary",
        "tags": [
          "bizzybee"
        ]
      }
    },
    "/api/v1/inventory/status": {
      "get": {
        "operationId": "inventory_status_api_v1_inventory_status_get",
        "parameters": [
          {
            "in": "query",
            "name": "question",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "minLength": 1,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Question"
            }
          },
          {
            "in": "query",
            "name": "language",
            "required": false,
            "schema": {
              "default": "en",
              "maxLength": 16,
              "minLength": 2,
              "title": "Language",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "window_days",
            "required": false,
            "schema": {
              "default": 7,
              "maximum": 90,
              "minimum": 1,
              "title": "Window Days",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "as_of",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "As Of"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Inventory Status",
        "tags": [
          "bizzybee"
        ]
      }
    },
    "/api/v1/me": {
      "get": {
        "operationId": "me_api_v1_me_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Me Api V1 Me Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Me",
        "tags": [
          "bizzybee"
        ]
      }
    },
    "/api/v1/query": {
      "post": {
        "operationId": "query_api_v1_query_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QueryRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QueryResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Query",
        "tags": [
          "bizzybee"
        ]
      }
    },
    "/api/v1/sales-inventory/alerts": {
      "get": {
        "operationId": "sales_inventory_alerts_api_v1_sales_inventory_alerts_get",
        "parameters": [
          {
            "in": "query",
            "name": "window_days",
            "required": false,
            "schema": {
              "default": 7,
              "maximum": 90,
              "minimum": 1,
              "title": "Window Days",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "as_of",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "As Of"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/BusinessAlert"
                  },
                  "title": "Response Sales Inventory Alerts Api V1 Sales Inventory Alerts Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Sales Inventory Alerts",
        "tags": [
          "bizzybee"
        ]
      }
    },
...
    "/api/v1/sales/summary": {
      "get": {
        "operationId": "sales_summary_api_v1_sales_summary_get",
        "parameters": [
          {
            "in": "query",
            "name": "question",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "minLength": 1,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Question"
            }
          },
          {
            "in": "query",
            "name": "language",
            "required": false,
            "schema": {
              "default": "en",
              "maxLength": 16,
              "minLength": 2,
              "title": "Language",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "window_days",
            "required": false,
            "schema": {
              "default": 7,
              "maximum": 90,
              "minimum": 1,
              "title": "Window Days",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "as_of",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "As Of"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Sales Summary",
        "tags": [
          "bizzybee"
        ]
      }
    },
...
  }
}